third_party/blink/web_tests/external/wpt/css/css-fonts/math-script-level-and-math-style/math-script-level-001.tentative-expected.txt

This is a testharness.js-based test.
FAIL Initial value of math-depth assert_equals: expected "0" but got ""
FAIL Inherited values of math-depth assert_equals: expected "11" but got ""
FAIL Specified math-depth: auto-add assert_equals: expected "10" but got ""
FAIL Specified math-depth: <integer> assert_equals: expected "11" but got ""
FAIL Specified math-depth: add(<integer>) assert_equals: expected "15" but got ""
Harness: the test ran to completion.
